refactor(certmanager): address P3 review findings
- Remove unused fields: Certificate.IPAddresses/URIs/Labels, Order.URL, Challenge.Token
- Collapse filterCertificatesByRBAC/filterIssuersByRBAC into generic filterByRBAC[T]
- Poller reuses Handler cache via CachedCertificates() with direct-list fallback
- Extract Discoverer to dedicated discovery.go file
- Simplify parseTimeField to use unstructured.NestedString directly
