+++ backend/internal/certmanager/handler.go
@@ -26,126 +26,13 @@ import (
 	"github.com/kubecenter/kubecenter/internal/server/middleware"
 )
 
-// ============================================================================
-// Discoverer
-// ============================================================================
-
 const (
-	staleDuration      = 5 * time.Minute
-	certManagerNS      = "cert-manager"
-	cacheTTL           = 30 * time.Second
-	maxRenewRetries    = 1
-	issuingCondType    = "Issuing"
-	issuingReason      = "ManuallyTriggered"
-	issuingMessage     = "Certificate re-issuance manually triggered"
+	cacheTTL        = 30 * time.Second
+	maxRenewRetries = 1
+	issuingCondType = "Issuing"
+	issuingReason   = "ManuallyTriggered"
+	issuingMessage  = "Certificate re-issuance manually triggered"
 )
-
-// Discoverer probes the cluster for cert-manager CRDs and maintains cached discovery state.
-type Discoverer struct {
-	k8sClient *k8s.ClientFactory
-	logger    *slog.Logger
-
-	mu     sync.RWMutex
-	status CertManagerStatus
-}
-
-// NewDiscoverer creates a new cert-manager discoverer.
-func NewDiscoverer(k8sClient *k8s.ClientFactory, logger *slog.Logger) *Discoverer {
-	return &Discoverer{
-		k8sClient: k8sClient,
-		logger:    logger,
-		status: CertManagerStatus{
-			LastChecked: time.Now().UTC(),
-		},
-	}
-}
-
-// Status returns a copy of the cached cert-manager status.
-// If the cache is stale (older than staleDuration), it triggers a re-probe.
-func (d *Discoverer) Status(ctx context.Context) CertManagerStatus {
-	d.mu.RLock()
-	if time.Since(d.status.LastChecked) < staleDuration {
-		status := d.status
-		d.mu.RUnlock()
-		return status
-	}
-	d.mu.RUnlock()
-
-	return d.Probe(ctx)
-}
-
-// Probe checks if cert-manager.io/v1 CRDs exist and updates cached state.
-func (d *Discoverer) Probe(ctx context.Context) CertManagerStatus {
-	d.mu.Lock()
-	defer d.mu.Unlock()
-
-	now := time.Now().UTC()
-	disco := d.k8sClient.DiscoveryClient()
-
-	status := CertManagerStatus{
-		LastChecked: now,
-	}
-
-	// Check for cert-manager CRDs
-	cmResources, err := disco.ServerResourcesForGroupVersion("cert-manager.io/v1")
-	if err != nil || cmResources == nil {
-		d.logger.Debug("cert-manager CRDs not found", "error", err)
-		d.status = status
-		return status
-	}
-
-	// Check if Certificate kind exists
-	hasCert := false
-	for _, r := range cmResources.APIResources {
-		if r.Kind == "Certificate" {
-			hasCert = true
-			break
-		}
-	}
-
-	if !hasCert {
-		d.logger.Debug("cert-manager Certificate CRD not found")
-		d.status = status
-		return status
-	}
-
-	status.Detected = true
-
-	// Probe the cert-manager namespace for deployment and version
-	cs := d.k8sClient.BaseClientset()
-	deps, err := cs.AppsV1().Deployments(certManagerNS).List(ctx, metav1.ListOptions{
-		LabelSelector: "app.kubernetes.io/name=cert-manager",
-	})
-	if err == nil && len(deps.Items) > 0 {
-		status.Namespace = certManagerNS
-		dep := deps.Items[0]
-		if v, ok := dep.Labels["app.kubernetes.io/version"]; ok {
-			status.Version = v
-		} else if len(dep.Spec.Template.Spec.Containers) > 0 {
-			img := dep.Spec.Template.Spec.Containers[0].Image
-			for i := len(img) - 1; i >= 0; i-- {
-				if img[i] == ':' {
-					status.Version = img[i+1:]
-					break
-				}
-			}
-		}
-	}
-
-	d.status = status
-	d.logger.Info("cert-manager discovery completed",
-		"detected", status.Detected,
-		"namespace", status.Namespace,
-		"version", status.Version,
-	)
-
-	return status
-}
-
-// IsAvailable returns true if cert-manager was detected.
-func (d *Discoverer) IsAvailable(ctx context.Context) bool {
-	return d.Status(ctx).Detected
-}
 
 // ============================================================================
 // Handler
@@ -260,38 +147,30 @@ func (h *Handler) auditLog(r *http.Request, user *auth.User, action audit.Action
 // RBAC filter helpers
 // ============================================================================
 
-// filterCertificatesByRBAC returns only certificates the user can access.
-func (h *Handler) filterCertificatesByRBAC(ctx context.Context, user *auth.User, items []Certificate) []Certificate {
-	allowed := make(map[string]bool)
-	result := make([]Certificate, 0, len(items))
-	for _, c := range items {
-		ok, checked := allowed[c.Namespace]
-		if !checked {
-			ok = h.canAccess(ctx, user, "get", "certificates", c.Namespace)
-			allowed[c.Namespace] = ok
-		}
-		if ok {
-			result = append(result, c)
-		}
-	}
-	return result
+// namespacedResource is implemented by types that carry a Kubernetes namespace.
+type namespacedResource interface {
+	getNamespace() string
 }
 
-// filterIssuersByRBAC returns only issuers the user can access.
-func (h *Handler) filterIssuersByRBAC(ctx context.Context, user *auth.User, items []Issuer) []Issuer {
-	allowed := make(map[string]bool)
-	result := make([]Issuer, 0, len(items))
-	for _, iss := range items {
-		ok, checked := allowed[iss.Namespace]
-		if !checked {
-			ok = h.canAccess(ctx, user, "get", "issuers", iss.Namespace)
-			allowed[iss.Namespace] = ok
+func (c Certificate) getNamespace() string { return c.Namespace }
+func (i Issuer) getNamespace() string      { return i.Namespace }
+
+// filterByRBAC returns only items the user can access in their respective namespaces.
+func filterByRBAC[T namespacedResource](ctx context.Context, h *Handler, user *auth.User, resource string, items []T) []T {
+	nsAllow := map[string]bool{}
+	out := make([]T, 0, len(items))
+	for _, item := range items {
+		ns := item.getNamespace()
+		allowed, ok := nsAllow[ns]
+		if !ok {
+			allowed = h.canAccess(ctx, user, "get", resource, ns)
+			nsAllow[ns] = allowed
 		}
-		if ok {
-			result = append(result, iss)
+		if allowed {
+			out = append(out, item)
 		}
 	}
-	return result
+	return out
 }
 
 // ============================================================================
@@ -391,6 +270,15 @@ func (h *Handler) fetchAll(ctx context.Context, gen uint64) (*cachedData, error)
 	return data, nil
 }
 
+// CachedCertificates returns the cached certificate list (for use by the Poller).
+func (h *Handler) CachedCertificates(ctx context.Context) ([]Certificate, error) {
+	data, err := h.getCached(ctx)
+	if err != nil {
+		return nil, err
+	}
+	return data.certificates, nil
+}
+
 // ============================================================================
 // Read handlers
 // ============================================================================
@@ -425,7 +313,7 @@ func (h *Handler) HandleListCertificates(w http.ResponseWriter, r *http.Request)
 		return
 	}
 
-	filtered := h.filterCertificatesByRBAC(r.Context(), user, data.certificates)
+	filtered := filterByRBAC(r.Context(), h, user, "certificates", data.certificates)
 
 	// Optional namespace filter
 	if ns := r.URL.Query().Get("namespace"); ns != "" {
@@ -591,7 +479,7 @@ func (h *Handler) HandleListIssuers(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	filtered := h.filterIssuersByRBAC(r.Context(), user, data.issuers)
+	filtered := filterByRBAC(r.Context(), h, user, "issuers", data.issuers)
 	httputil.WriteData(w, filtered)
 }
 
@@ -643,7 +531,7 @@ func (h *Handler) HandleListExpiring(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	certs := h.filterCertificatesByRBAC(r.Context(), user, data.certificates)
+	certs := filterByRBAC(r.Context(), h, user, "certificates", data.certificates)
 
 	expiring := make([]ExpiringCertificate, 0)
 	for _, c := range certs {
